Merged revisions 162264-162265,162273 via svnmerge from file:///srv/subversion/repos/asterisk/branches/1.4
........
r162264 | murf | 2008-12-09 14:20:54 -0600 (Tue, 09 Dec 2008) | 1 line

In discussion with seanbright on #asterisk-dev,  I have added a default rule, and an option to suppress the default rule from being generated in the flex output, for the sake of those OS's where they didn't tweak flex's ECHO macro, and the compiler doesn't like it. The regressions are OK with this.
........
r162265 | mmichelson | 2008-12-09 14:28:44 -0600 (Tue, 09 Dec 2008) | 6 lines

If we fail to start a thread for the pbx to run in, we need to
be sure to decrease the number of active calls on the system.

This fix may relate to ABE-1713, but it is not certain yet.


........
r162273 | file | 2008-12-09 14:44:32 -0600 (Tue, 09 Dec 2008) | 4 lines

Fix double declaration of 'x' on the PPC platform.
(closes issue #14038)
Reported by: ffloimair

........