Merged revisions 162997,163037 via svnmerge from file:///srv/subversion/repos/asterisk/trunk
........
r162997 | file | 2008-12-11 09:05:49 -0600 (Thu, 11 Dec 2008) | 4 lines

When a device registers to use it is entirely possible that they may be in use, so tell the core that we don't know the devstate and have it ask us for it.
(closes issue #13525)
Reported by: pj

........
r163037 | seanbright | 2008-12-11 09:40:07 -0600 (Thu, 11 Dec 2008) | 8 lines

Fix some of the grammar issues in doc/tex/qos.tex.

(closes issue #14049)
Reported by: kshumard
Patches:
      doc.tex.qos.tex.patch uploaded by kshumard (license 92)
      (Slight modifications by seanbright)

........