Merged revisions 163166,163168,163171 via svnmerge from file:///srv/subversion/repos/asterisk/trunk
........
r163166 | mmichelson | 2008-12-11 13:40:18 -0600 (Thu, 11 Dec 2008) | 3 lines

Reduce indentation level of ast_feature_request_and_dial


........
r163168 | tilghman | 2008-12-11 14:02:35 -0600 (Thu, 11 Dec 2008) | 5 lines

Sometimes even Linux needs -lm to link libtonezone, such as when libtonezone
is compiled statically.
(closes issue #13887)
 Reported by: tzafrir

........
r163171 | russell | 2008-12-11 14:07:47 -0600 (Thu, 11 Dec 2008) | 16 lines

Fix the "failed" extension for outgoing calls.  

The conversion to use ast_check_hangup() everywhere instead of checking the softhangup
flag directly introduced this problem.  The issue is that ast_check_hangup() checked
for tech_pvt to be NULL.  Unfortunately, this will be NULL is some valid circumstances,
such as with a dummy channel.

The fix is simple.  Don't check tech_pvt.  It's pointless, because the code path that
sets this to NULL is when the channel hangup callback gets called.  This happens inside
of ast_hangup(), which is the same function responsible for freeing the channel.  Any
code calling ast_check_hangup() better not be calling it after that point, and if so,
we have a bigger problem at hand.

(closes issue #14035)
Reported by: erogoza

........