A good place to checkpoint, maybe. A ton of now useless debug.
I was getting stuck in the taskprocessor task execution, 
because the dummy channel creation needed to init the lock; 
the variable_helper func sets that lock.

And, last night as I was retiring for the night,
I realized that taskprocessor is not an appropriate
mechanism for this sort of usage; but it might
be for others.

But it all proves that taskprocessor works great
with multiple task engines.

The sqlite3 loading was subscribing to events way
too early; fixed that.

I need to correct formatting and get rid of the
debug. Stay tuned.