Merged revisions 164976,164978 via svnmerge from file:///srv/subversion/repos/asterisk/trunk
................
r164976 | tilghman | 2008-12-16 16:57:17 -0600 (Tue, 16 Dec 2008) | 7 lines

Add timezone to the possible fields in a timespec.
(closes issue #14028)
 Reported by: mostyn
 Patches: 
       timezone-v2.patch uploaded by mostyn (license 398)
       (with additional code guideline fixes and a memory leak fix by me - license 14)

................
r164978 | mmichelson | 2008-12-16 17:06:04 -0600 (Tue, 16 Dec 2008) | 15 lines

Merged revisions 164977 via svnmerge from 
https://origsvn.digium.com/svn/asterisk/branches/1.4

........
r164977 | mmichelson | 2008-12-16 17:04:27 -0600 (Tue, 16 Dec 2008) | 7 lines

After looking through SIP registration code most of the day, this
is one of the few things I could find that was just plain wrong.
Even though it probably isn't possible for it to happen, it seems weird
to have code that checks if a pointer is NULL and then immediately dereferences
that pointer if it was NULL.


........

................