Merged revisions 165142-165143,165145 via svnmerge from file:///srv/subversion/repos/asterisk/trunk
........
r165142 | mmichelson | 2008-12-17 11:52:50 -0600 (Wed, 17 Dec 2008) | 10 lines

Use the create_vm_state_from_user function in a place where
it was not being used before. Also, I've moved the urgent
folder check in messagecount() up a bit so that the flow is
a bit better.

This was something I noticed while taking a look at issue
#13973, although I don't think this is the underlying cause
of the issue.


........
r165143 | mmichelson | 2008-12-17 11:53:37 -0600 (Wed, 17 Dec 2008) | 3 lines

And actually assign the function to a pointer...


........
r165145 | russell | 2008-12-17 11:56:25 -0600 (Wed, 17 Dec 2008) | 2 lines

argsep is used as an attribute for an argument, as well

........