Merged revisions 165469 via svnmerge from file:///srv/subversion/repos/asterisk/trunk
........
r165469 | russell | 2008-12-18 08:23:22 -0600 (Thu, 18 Dec 2008) | 6 lines

Add a \todo note for app_originate.

Jared Smith suggested that we add a way to be able to set variables
and functions on the outbound channel.  I think that it's a great idea, so I
have added it as a todo so that it gets done at some point.

........