Merged revisions 165537,165591 via svnmerge from file:///srv/subversion/repos/asterisk/branches/1.4
........
r165537 | file | 2008-12-18 10:07:37 -0600 (Thu, 18 Dec 2008) | 6 lines

Do not crash if we are not passed in a followme id.
(closes issue #14106)
Reported by: ys
Patches:
      app_followme.c.2.diff uploaded by ys (license 281)

........
r165591 | file | 2008-12-18 11:11:42 -0600 (Thu, 18 Dec 2008) | 4 lines

Only care about a compatible codec for early bridging if we are actually bridging to another channel. If we are not we actually want to bring the audio back to us.
(closes issue #13545)
Reported by: davidw

........