Some changes to the CLI commands involved in reloading queues
It was pointed out to me on IRC that some of the CLI commands I have
added violate the typical "module verb object" order of things, so
I have decided to redo the CLI commands from scratch. It's not very
difficult to do, so I don't mind. This is the first commit to that
end. Now there is a single "queue reload" command which may be used
to reload any combination of queue members, parameters, stats, and rules
for all queues or for a single queue. The idea is also to introduce
one or multiple CLI aliases so that there may be multiple methods
to execute a single logical path.

Introducing CLI aliases is probably also going to require me to make
the name of the queue the final argument for these CLI commands. For
now, the next step is to write the tab-completion function