Another earth-shattering change to the CLI commands.
I had a discussion with Kevin about this branch last
week, and he gave some wonderful suggestions for the
new CLI commands offered in this branch. Essentially,
I had made a mire of a command that was trying to do
way too much for not much benefit. We came to an
agreement that two changes needed to be made.

1) Simplify the 'queue reload' command significantly
2) Allow for multiple queues to be specified for the
   new CLI commands.

This commit accomplishes part 1. This changes the queue
reload command to not allow arbitrary combinations of
qualifiers to be used when reloading queues. This simplifies
the command both from a user interface perspective, as
well as a coding perspective, so it is definitely a
good thing.

This code passes the "it compiles" test, but I have not
actually attempted to use the CLI commands at all yet
as I would like to finish part 2 of the modifications
first.