Creating a new branch to fix issue #14220
The problem arose when chan_sip converted its sip_request to use
an expandable ast_str structure instead of using a statically-sized
buffer.

A sip_request structure maintains two arrays of character pointers, called
header and line. These point to each individual header or SDP line in the
SIP request. The problem that this branch will be addressing is that when
an ast_str expands to hold more data, the header and line pointers are then
pointing to invalid memory.

The plan of action here is to change the header and line arrays to contain
offsets into the ast_str struct instead character pointers. This way, if the
ast_str is reallocated to allow for more data to be stored inside, the same
offsets can be applied to the newly allocated area and all will be well.