Merged revisions 170393,170457,170460,170463 via svnmerge from https://origsvn.digium.com/svn/asterisk/trunk
................
r170393 | mmichelson | 2009-01-23 08:44:27 -0700 (Fri, 23 Jan 2009) | 36 lines

Merged revisions 170392 via svnmerge from 
https://origsvn.digium.com/svn/asterisk/branches/1.4

........
r170392 | mmichelson | 2009-01-23 09:40:39 -0600 (Fri, 23 Jan 2009) | 28 lines

Fix broken call pickup

There was a subtle change in ast_do_masquerade which
resulted in failed attempts to pickup calls. The problem
was that the value of the AST_FLAG_OUTGOING flag was
copied from the clone to the original channel. In the case
of call pickup, this meant that the AST_FLAG_OUTGOING flag
ended up being cleared on the channel that was attempting
to execute the pickup.

Because this flag was not set, when ast_read came across
an answer frame, it ignored it. The result of this was that
the calling channel was never properly answered.

This fix changes the behavior in ast_do_masquerade to set
the flags on the original channel to the union of the flags
on the clone channel. This way, if the AST_FLAG_OUTGOING
flag is set on either of the two channels involved in the
masquerade, the resulting channel will have the flag set
as well.

(closes issue #14206)
Reported by: francesco_r
Patches:
      14206.patch uploaded by putnopvut (license 60)
Tested by: francesco_r, aragon, putnopvut


........

................
r170457 | dbailey | 2009-01-23 09:35:30 -0700 (Fri, 23 Jan 2009) | 5 lines

MWI messages included in CID spill was not being properly handled and prevented the call from being processed
(issue #14313)
Reported by: seandarcy
Tested by: dbailey

................
r170460 | mvanbaak | 2009-01-23 09:57:07 -0700 (Fri, 23 Jan 2009) | 8 lines

Dont clear the display of skinny phones when not needed.

(closes issue #13182)
Reported by: pj
Patches:
      2009011901_dontcleardisplay.diff.txt uploaded by mvanbaak (license 7)
Tested by: mvanbaak, pj

................
r170463 | mattf | 2009-01-23 10:03:41 -0700 (Fri, 23 Jan 2009) | 1 line

We should not do restart messages if we're in PTMP mode
................