Fix a race condition in app_confbridge
In order for this to occur, the option to play music on hold
if there is only one person in the conference must be set.
If a second caller entered the conference while the first
caller was listening to the "conf-onlyperson" prompt, then
the first caller would have music on hold played to him even
though a second caller had entered the conference. This was due
to the fact that playing the prompt to the first caller caused
the conference bridge to become unlocked, thus allowing the second
caller to slip in.

The fix for this is to re-check the number of users after the prompt
completes.