Initial change of app_chanspy to use autochan
This is an "it compiles" type checkin. I have not looked very 
closely at app_chanspy to see if more can be removed as a result
of this change, nor have I run any tests to be sure that everything
still works as it should.

One thing I know is off here is that app_chanspy follows the exact
opposite of the locking order described in autochan.h so that should
be addressed.