Add autochan interface to ast_channel_ao2 and convert chanspy and mixmonitor to use it
chanspy and mixmonitor are examples of applications which attempt to keep a pointer to
a channel in a thread external to the channel's main (i.e. PBX) thread. In both cases,
the channel that they reference may change due to a masquerade. The autochan interface
is a method by which a pointer to a channel may be automatically updated if a channel
is masqueraded.

This also finishes the conversion of app_chanspy and app_mixmonitor to use refcounted
channels.