Branch for backporting (evil art form, indeed) timer_b code from trunk, also implementing some sort of dialplan support for limiting the number of seconds a call can hang in ring(AST) or call(SIP) state.
The best solution would be to implement an option to Dial() that
works across all channel types, but I don't know if the budget for
this SIP project allows that or if I can pull that trick out of
my head at all. Diving down into dial and bridge is, well, not my
piece of cake. SIP on the other hand, is exactly my cup of tea.

Witty, witty. Time to start coding.